refactor: DeleteHistory를 common 패키지로 변경
