+++ src/main/java/nextstep/common/domain/DeleteHistory.java
@@ -1,5 +1,6 @@
-package nextstep.qna.domain;
+package nextstep.common.domain;
 
+import nextstep.qna.domain.ContentType;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
