+++ src/main/java/nextstep/qna/domain/Answers.java
@@ -1,13 +1,12 @@
 package nextstep.qna.domain;
 
+import nextstep.common.domain.DeleteHistory;
+import nextstep.common.domain.DeleteHistoryTargets;
 import nextstep.qna.CannotDeleteException;
-import nextstep.qna.CannotConvertException;
 import nextstep.users.domain.NsUser;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Answers {
