+++ src/test/java/nextstep/qna/domain/AnswerTest.java
@@ -1,5 +1,6 @@
 package nextstep.qna.domain;
 
+import nextstep.common.domain.DeleteHistory;
 import nextstep.users.domain.NsUserTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
