+++ src/test/java/nextstep/qna/domain/AnswersTest.java
@@ -1,5 +1,6 @@
 package nextstep.qna.domain;
 
+import nextstep.common.domain.DeleteHistory;
 import nextstep.qna.CannotDeleteException;
 import nextstep.users.domain.NsUserTest;
 import org.junit.jupiter.api.BeforeEach;
